ats_backend/requirements.txt
fastapi
uvicorn
pytesseract
pillow
pdf2image
python-docx
PyPDF2
sqlalchemy
sentence-transformers
torch
